+++ src/ScottPlot4/ScottPlot/Palettes/Category10.cs
@@ -47,10 +47,13 @@ OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
-namespace ScottPlot.Drawing.Colorsets
+namespace ScottPlot.Palettes;
+
+public class Category10 : HexPaletteBase, IPalette
 {
-    public class Category10 : HexColorset, IPalette
+    internal override string[] HexColors => new string[]
     {
-        public override string[] hexColors => Common.HexPalettes.Category10.Colors;
-    }
+        "#1f77b4", "#ff7f0e", "#2ca02c", "#d62728", "#9467bd",
+        "#8c564b", "#e377c2", "#7f7f7f", "#bcbd22", "#17becf",
+    };
 }