+++ src/ScottPlot4/ScottPlot/Palettes/DarkPastel.cs
@@ -2,10 +2,14 @@
  * © Cynthia Brewer, Mark Harrower and The Pennsylvania State University
  * This palette is the lighter-color version of the 'Dark' palette.
  */
-namespace ScottPlot.Drawing.Colorsets
+
+namespace ScottPlot.Palettes;
+
+public class DarkPastel : HexPaletteBase, IPalette
 {
-    public class DarkPastel : HexColorset, IPalette
+    internal override string[] HexColors => new string[]
     {
-        public override string[] hexColors => Common.HexPalettes.DarkPastel.Colors;
-    }
+        "#66c2a5", "#fc8d62", "#8da0cb", "#e78ac3", "#a6d854",
+        "#ffd92f", "#e5c494", "#b3b3b3",
+    };
 }