+++ src/ScottPlot4/ScottPlot/Palettes/OneHalfDark.cs
@@ -1,12 +1,13 @@
 ﻿/* Sourced from Son A. Pham's Sublime color scheme by the same name
  * https://github.com/sonph/onehalf
  */
-namespace ScottPlot.Drawing.Colorsets
-{
-    public class OneHalfDark : HexColorset, IPalette
-    {
-        // suggested background: #2e3440
-        public override string[] hexColors => Common.HexPalettes.OneHalfDark.Colors;
 
-    }
+namespace ScottPlot.Palettes;
+
+public class OneHalfDark : HexPaletteBase, IPalette
+{
+    internal override string[] HexColors => new string[]
+    {
+        "#e06c75", "#98c379", "#e5c07b", "#61aff0", "#c678dd", "#56b6c2", "#dcdfe4"
+    };
 }