+++ src/ScottPlot4/ScottPlot/Palettes/Penumbra.cs
@@ -1,6 +1,14 @@
-﻿namespace ScottPlot.Drawing.Colorsets;
+﻿/* Sourced from Neal McKee's Penumbra color theme:
+ * https://github.com/nealmckee/penumbra/blob/main/penumbra.tsv
+ * https://github.com/nealmckee/penumbra#accent-colour-palettes-2
+ */
 
-public class Penumbra : HexColorset, IPalette
+namespace ScottPlot.Palettes;
+
+public class Penumbra : HexPaletteBase, IPalette
 {
-    public override string[] hexColors => Common.HexPalettes.Penumbra.Colors;
+    internal override string[] HexColors => new string[]
+    {
+        "#CB7459", "#A38F2D", "#46A473", "#00A0BE", "#7E87D6", "#BD72A8"
+    };
 }