+++ src/ScottPlot4/ScottPlot/Palettes/Tsitsulin.cs
@@ -2,10 +2,19 @@
  * http://tsitsul.in/blog/coloropt
  * https://github.com/xgfs/coloropt
  */
-namespace ScottPlot.Drawing.Colorsets
+
+namespace ScottPlot.Palettes;
+
+public class Tsitsulin : HexPaletteBase, IPalette
 {
-    public class Tsitsulin : HexColorset, IPalette
+    internal override string[] HexColors => new string[]
     {
-        public override string[] hexColors => Common.HexPalettes.Tsitsulin.Colors;
-    }
+        "#ebac23", "#b80058", "#008cf9", "#006e00", "#00bbad",
+        "#d163e6", "#b24502", "#ff9287", "#5954d6", "#00c6f8",
+        "#878500", "#00a76c",
+        "#f6da9c", "#ff5caa", "#8accff", "#4bff4b", "#6efff4",
+        "#edc1f5", "#feae7c", "#ffc8c3", "#bdbbef", "#bdf2ff",
+        "#fffc43", "#65ffc8",
+        "#aaaaaa",
+    };
 }